simplify commits
