+++ rampwf/hyperopt/tests/interfaces/header_in_files/classifier_kit/requirements.txt
@@ -1,0 +1,5 @@
+numpy
+scikit-learn
+pandas
+matplotlib
+seaborn
