+++ rampwf/hyperopt/tests/interfaces/header_in_files/classifier_kit/submissions/starting_kit_titanic_hyperopt/hyperopt_output/summary.csv
@@ -1,0 +1,9 @@
+complex_features,imputer_strategy,logreg_C,n_folds,train_auc_m,train_acc_m,train_nll_m,valid_auc_m,valid_acc_m,valid_nll_m,train_time_m,valid_time_m,n_train_m,n_valid_m,train_auc_s,train_acc_s,train_nll_s,valid_auc_s,valid_acc_s,valid_nll_s,train_time_s,valid_time_s,n_train_s,n_valid_s
+False,mean,0.1,8,0.8456815203145478,0.8006161971830986,0.4923561720531346,0.8557731331168831,0.8107638888888888,0.4891622330538865,0.020332008600234985,0.0021424293518066406,284.0,72.0,0.010136242427118549,0.0056267246783695025,0.009700610657773834,0.04631753977308916,0.028692767587261067,0.029470208620682364,0.0006822508810130299,0.0001602828571209061,0.0,0.0
+False,mean,1.0,8,0.8534338138925295,0.8098591549295775,0.4513273502763074,0.8624695616883117,0.8072916666666666,0.44186752802883167,0.020251095294952393,0.0021677613258361816,284.0,72.0,0.011318520097412722,0.008827931979267711,0.013534127638829729,0.0475520701332065,0.027214021278462727,0.04613278156418855,0.0006686987752690287,0.00018069053441161216,0.0,0.0
+False,median,0.1,8,0.8456815203145478,0.8006161971830986,0.4923561720531346,0.8557731331168831,0.8107638888888888,0.4891622330538865,0.020137906074523926,0.002192378044128418,284.0,72.0,0.010136242427118549,0.0056267246783695025,0.009700610657773834,0.04631753977308916,0.028692767587261067,0.029470208620682364,0.0006048956777728816,0.000150155040097624,0.0,0.0
+False,median,0.9,8,0.8528440366972477,0.8089788732394366,0.4520242109361346,0.8624695616883117,0.8072916666666666,0.44232979092544333,0.020824790000915527,0.00211217999458313,284.0,72.0,0.01175233509776517,0.00955072311169395,0.013776629680008244,0.04669724733727432,0.031000992047618532,0.045703805523972275,0.0004790343175184534,4.82610152276858e-05,0.0,0.0
+True,mean,0.1,8,0.8652031454783748,0.8142605633802817,0.4514756731154752,0.8665280032467533,0.8055555555555556,0.4516098360887191,0.024923652410507202,0.005901366472244263,284.0,72.0,0.013236958256024396,0.010091751834878699,0.01386345552468407,0.04589215733373116,0.02347650707579214,0.04543807691139413,0.0016420642777369078,0.00019293981294119837,0.0,0.0
+True,mean,1.0,8,0.8682306684141546,0.818661971830986,0.43294555394406037,0.8728185876623377,0.8090277777777778,0.42957920975589003,0.02385982871055603,0.006170153617858887,284.0,72.0,0.012335515624466482,0.009596973857313304,0.015784925049280182,0.0508090475238741,0.02946278254943947,0.058875519566498814,0.0007112071365381468,0.0005544055710511826,0.0,0.0
+True,median,0.01,8,0.8271625163826999,0.7544014084507042,0.5236636360201472,0.8218851461038961,0.7309027777777778,0.5327518109703279,0.019297003746032715,0.006231069564819336,284.0,72.0,0.012173654343358,0.011716111367054442,0.009235246913456166,0.046697247337274365,0.03144230950742158,0.03158090557342812,0.0015210502668520135,0.0012843259117325426,0.0,0.0
+True,median,0.1,8,0.8652031454783748,0.8142605633802817,0.4514756731154752,0.8665280032467533,0.8055555555555556,0.4516098360887191,0.024460822343826294,0.005753844976425171,284.0,72.0,0.013236958256024396,0.010091751834878699,0.01386345552468407,0.04589215733373116,0.02347650707579214,0.04543807691139413,0.0005600876079264603,0.00013992399333909245,0.0,0.0
