--- rampwf/hyperopt/tests/interfaces/header_in_files/generative_classifier/data/click/dtypes.json
@@ -1,1 +1,0 @@
-{"impression": "num", "url_hash": "num", "advertiser_id": "num", "depth": "num", "position": "num", "ad_id": "num", "query_id": "num", "keyword_id": "num", "title_id": "num", "description_id": "num", "user_id": "num"}
