--- rampwf/hyperopt/tests/interfaces/header_in_files/generative_classifier/data/cover_type/dtypes.json
@@ -1,56 +1,0 @@
-{
-  "Elevation": "num",
-  "Aspect" : "num",
-  "Horizontal_Distance_To_Hydrology" : "num",
-  "Vertical_Distance_To_Hydrology" :  "num",
-  "Horizontal_Distance_To_Roadways" : "num",
-  "Hillshade_9am" : "num",
-  "Hillshade_Noon" :  "num",
-  "Hillshade_3pm" :  "num",
-  "Horizontal_Distance_To_Fire_Points" :  "num",
-  "Slope" : "num",
-  "Wilderness_Area1" : "cat",
-  "Wilderness_Area2" :  "cat",
-  "Wilderness_Area3" : "cat",
-  "Wilderness_Area4" :  "cat",
-  "Soil_Type1" : "cat",
-  "Soil_Type2" : "cat",
-  "Soil_Type3" : "cat",
-  "Soil_Type4" : "cat",
-  "Soil_Type5" : "cat",
-  "Soil_Type6" : "cat",
-  "Soil_Type7" : "cat",
-  "Soil_Type8" : "cat",
-  "Soil_Type9" : "cat",
-  "Soil_Type10" : "cat",
-  "Soil_Type11" : "cat",
-  "Soil_Type12" : "cat",
-  "Soil_Type13" : "cat",
-  "Soil_Type14" : "cat",
-  "Soil_Type15" : "cat",
-  "Soil_Type16" : "cat",
-  "Soil_Type17" : "cat",
-  "Soil_Type18" : "cat",
-  "Soil_Type19" : "cat",
-  "Soil_Type20" : "cat",
-  "Soil_Type21" : "cat",
-  "Soil_Type22" : "cat",
-  "Soil_Type23" : "cat",
-  "Soil_Type24" : "cat",
-  "Soil_Type25" : "cat",
-  "Soil_Type26" : "cat",
-  "Soil_Type27" : "cat",
-  "Soil_Type28" : "cat",
-  "Soil_Type29" : "cat",
-  "Soil_Type30" : "cat",
-  "Soil_Type31" : "cat",
-  "Soil_Type32" : "cat",
-  "Soil_Type33" : "cat",
-  "Soil_Type34" : "cat",
-  "Soil_Type35" : "cat",
-  "Soil_Type36" : "cat",
-  "Soil_Type37" : "cat",
-  "Soil_Type38" : "cat",
-  "Soil_Type39" : "cat",
-  "Soil_Type40" : "cat"
-}
