--- rampwf/hyperopt/tests/interfaces/header_in_files/generative_classifier/submissions/rf_cover_type_hyperopt/hyperopt_output/summary.csv
@@ -1,5 +1,0 @@
-max_leaf_nodes,n_folds,train_acc_m,valid_acc_m,train_time_m,valid_time_m,n_train_m,n_valid_m,train_acc_s,valid_acc_s,train_time_s,valid_time_s,n_train_s,n_valid_s
-2.0,16,0.7566666666666667,0.77,0.0871461033821106,0.010737031698226929,900.0,100.0,0.0,0.0,0.006160739969357132,0.0002092375155958668,0.0,0.0
-4.0,16,0.7644444444444445,0.77,0.09188172221183777,0.011621236801147461,900.0,100.0,0.0,0.0,0.004860457487414362,0.0004727999486858881,0.0,0.0
-8.0,16,0.7766666666666666,0.77,0.09322832524776459,0.012001529335975647,900.0,100.0,0.0,0.0,0.003248169933851373,0.00015884419541710906,0.0,0.0
-16.0,16,0.7911111111111111,0.76,0.09621994197368622,0.012489110231399536,900.0,100.0,0.0,0.0,0.0030927279041452673,0.00017276069464297475,0.0,0.0
