--- rampwf/hyperopt/tests/interfaces/header_in_files/titanic/README.md
@@ -1,23 +1,0 @@
-# RAMP starting kit on the Titanic dataset
-
-Authors: Alexandre Gramfort & Balazs Kegl
-
-[![Build Status](https://travis-ci.org/ramp-kits/titanic.svg?branch=master)](https://travis-ci.org/ramp-kits/titanic)
-
-Go to [`ramp-worflow`](https://github.com/paris-saclay-cds/ramp-workflow) for more help on the [RAMP](http:www.ramp.studio) ecosystem.
-
-Install ramp-workflow (rampwf), then execute
-
-```
-ramp_test_submission
-```
-
-to test the starting kit submission (`submissions/starting_kit`) and
-
-```
-ramp_test_submission --submission=random_forest_20_5
-```
-
-to test `random_forest_20_5` or any other submission in `submissions`.
-
-Get started on this RAMP with the [dedicated notebook](titanic_starting_kit.ipynb).
