--- rampwf/hyperopt/tests/interfaces/header_in_files/titanic/data/test.csv
@@ -1,90 +1,0 @@
-PassengerId,Survived,Pclass,Name,Sex,Age,SibSp,Parch,Ticket,Fare,Cabin,Embarked
-270,1,1,"Bissette, Miss. Amelia",female,35.0,0,0,PC 17760,135.6333,C99,S
-23,1,3,"McGowan, Miss. Anna ""Annie""",female,15.0,0,0,330923,8.0292,,Q
-827,0,3,"Lam, Mr. Len",male,,0,0,1601,56.4958,,S
-397,0,3,"Olsson, Miss. Elina",female,31.0,0,0,350407,7.8542,,S
-129,1,3,"Peter, Miss. Anna",female,,1,1,2668,22.3583,F E69,C
-519,1,2,"Angle, Mrs. William A (Florence ""Mary"" Agnes Hughes)",female,36.0,1,0,226875,26.0,,S
-155,0,3,"Olsen, Mr. Ole Martin",male,,0,0,Fa 265302,7.3125,,S
-278,0,2,"Parkes, Mr. Francis ""Frank""",male,,0,0,239853,0.0,,S
-738,1,1,"Lesurer, Mr. Gustave J",male,35.0,0,0,PC 17755,512.3292,B101,C
-251,0,3,"Reed, Mr. James George",male,,0,0,362316,7.25,,S
-549,0,3,"Goldsmith, Mr. Frank John",male,33.0,1,1,363291,20.525,,S
-173,1,3,"Johnson, Miss. Eleanor Ileen",female,1.0,1,1,347742,11.1333,,S
-291,1,1,"Barber, Miss. Ellen ""Nellie""",female,26.0,0,0,19877,78.85,,S
-639,0,3,"Panula, Mrs. Juha (Maria Emilia Ojala)",female,41.0,0,5,3101295,39.6875,,S
-399,0,2,"Pain, Dr. Alfred",male,23.0,0,0,244278,10.5,,S
-326,1,1,"Young, Miss. Marie Grice",female,36.0,0,0,PC 17760,135.6333,C32,C
-219,1,1,"Bazzani, Miss. Albina",female,32.0,0,0,11813,76.2917,D15,C
-835,0,3,"Allum, Mr. Owen George",male,18.0,0,0,2223,8.3,,S
-764,1,1,"Carter, Mrs. William Ernest (Lucile Polk)",female,36.0,1,2,113760,120.0,B96 B98,S
-313,0,2,"Lahtinen, Mrs. William (Anna Sylfven)",female,26.0,1,1,250651,26.0,,S
-838,0,3,"Sirota, Mr. Maurice",male,,0,0,392092,8.05,,S
-228,0,3,"Lovell, Mr. John Hall (""Henry"")",male,20.5,0,0,A/5 21173,7.25,,S
-271,0,1,"Cairns, Mr. Alexander",male,,0,0,113798,31.0,,S
-473,1,2,"West, Mrs. Edwy Arthur (Ada Mary Worth)",female,33.0,1,2,C.A. 34651,27.75,,S
-126,1,3,"Nicola-Yarred, Master. Elias",male,12.0,1,0,2651,11.2417,,C
-37,1,3,"Mamee, Mr. Hanna",male,,0,0,2677,7.2292,,C
-532,0,3,"Toufik, Mr. Nakli",male,,0,0,2641,7.2292,,C
-638,0,2,"Collyer, Mr. Harvey",male,31.0,1,1,C.A. 31921,26.25,,S
-259,1,1,"Ward, Miss. Anna",female,35.0,0,0,PC 17755,512.3292,,C
-267,0,3,"Panula, Mr. Ernesti Arvid",male,16.0,4,1,3101295,39.6875,,S
-350,0,3,"Dimic, Mr. Jovan",male,42.0,0,0,315088,8.6625,,S
-169,0,1,"Baumann, Mr. John D",male,,0,0,PC 17318,25.925,,S
-754,0,3,"Jonkoff, Mr. Lalio",male,23.0,0,0,349204,7.8958,,S
-402,0,3,"Adams, Mr. John",male,26.0,0,0,341826,8.05,,S
-75,1,3,"Bing, Mr. Lee",male,32.0,0,0,1601,56.4958,,S
-117,0,3,"Connors, Mr. Patrick",male,70.5,0,0,370369,7.75,,Q
-746,0,1,"Crosby, Capt. Edward Gifford",male,70.0,1,1,WE/P 5735,71.0,B22,S
-843,1,1,"Serepeca, Miss. Augusta",female,30.0,0,0,113798,31.0,,C
-77,0,3,"Staneff, Mr. Ivan",male,,0,0,349208,7.8958,,S
-645,1,3,"Baclini, Miss. Eugenie",female,0.75,2,1,2666,19.2583,,C
-643,0,3,"Skoog, Miss. Margit Elizabeth",female,2.0,3,2,347088,27.9,,S
-422,0,3,"Charters, Mr. David",male,21.0,0,0,A/5. 13032,7.7333,,Q
-295,0,3,"Mineff, Mr. Ivan",male,24.0,0,0,349233,7.8958,,S
-2,1,1,"Cumings, Mrs. John Bradley (Florence Briggs Thayer)",female,38.0,1,0,PC 17599,71.2833,C85,C
-181,0,3,"Sage, Miss. Constance Gladys",female,,8,2,CA. 2343,69.55,,S
-177,0,3,"Lefebre, Master. Henry Forbes",male,,3,1,4133,25.4667,,S
-823,0,1,"Reuchlin, Jonkheer. John George",male,38.0,0,0,19972,0.0,,S
-206,0,3,"Strom, Miss. Telma Matilda",female,2.0,0,1,347054,10.4625,G6,S
-672,0,1,"Davidson, Mr. Thornton",male,31.0,1,0,F.C. 12750,52.0,B71,S
-174,0,3,"Sivola, Mr. Antti Wilhelm",male,21.0,0,0,STON/O 2. 3101280,7.925,,S
-571,1,2,"Harris, Mr. George",male,62.0,0,0,S.W./PP 752,10.5,,S
-531,1,2,"Quick, Miss. Phyllis May",female,2.0,1,1,26360,26.0,,S
-453,0,1,"Foreman, Mr. Benjamin Laventall",male,30.0,0,0,113051,27.75,C111,C
-88,0,3,"Slocovski, Mr. Selman Francis",male,,0,0,SOTON/OQ 392086,8.05,,S
-528,0,1,"Farthing, Mr. John",male,,0,0,PC 17483,221.7792,C95,S
-71,0,2,"Jenkin, Mr. Stephen Curnow",male,32.0,0,0,C.A. 33111,10.5,,S
-367,1,1,"Warren, Mrs. Frank Manley (Anna Sophia Atkinson)",female,60.0,1,0,110813,75.25,D37,C
-884,0,2,"Banfield, Mr. Frederick James",male,28.0,0,0,C.A./SOTON 34068,10.5,,S
-834,0,3,"Augustsson, Mr. Albert",male,23.0,0,0,347468,7.8542,,S
-208,1,3,"Albimona, Mr. Nassef Cassem",male,26.0,0,0,2699,18.7875,,C
-79,1,2,"Caldwell, Master. Alden Gates",male,0.83,0,2,248738,29.0,,S
-865,0,2,"Gill, Mr. John William",male,24.0,0,0,233866,13.0,,S
-379,0,3,"Betros, Mr. Tannous",male,20.0,0,0,2648,4.0125,,C
-5,0,3,"Allen, Mr. William Henry",male,35.0,0,0,373450,8.05,,S
-613,1,3,"Murphy, Miss. Margaret Jane",female,,1,0,367230,15.5,,Q
-386,0,2,"Davies, Mr. Charles Henry",male,18.0,0,0,S.O.C. 14879,73.5,,S
-517,1,2,"Lemore, Mrs. (Amelia Milley)",female,34.0,0,0,C.A. 34260,10.5,F33,S
-533,0,3,"Elias, Mr. Joseph Jr",male,17.0,1,1,2690,7.2292,,C
-765,0,3,"Eklund, Mr. Hans Linus",male,16.0,0,0,347074,7.775,,S
-523,0,3,"Lahoud, Mr. Sarkis",male,,0,0,2624,7.225,,C
-325,0,3,"Sage, Mr. George John Jr",male,,8,2,CA. 2343,69.55,,S
-187,1,3,"O'Brien, Mrs. Thomas (Johanna ""Hannah"" Godfrey)",female,,1,0,370365,15.5,,Q
-618,0,3,"Lobb, Mrs. William Arthur (Cordelia K Stanlick)",female,26.0,1,0,A/5. 3336,16.1,,S
-619,1,2,"Becker, Miss. Marion Louise",female,4.0,2,1,230136,39.0,F4,S
-305,0,3,"Williams, Mr. Howard Hugh ""Harry""",male,,0,0,A/5 2466,8.05,,S
-633,1,1,"Stahelin-Maeglin, Dr. Max",male,32.0,0,0,13214,30.5,B50,C
-612,0,3,"Jardin, Mr. Jose Neto",male,,0,0,SOTON/O.Q. 3101305,7.05,,S
-721,1,2,"Harper, Miss. Annie Jessie ""Nina""",female,6.0,0,1,248727,33.0,,S
-526,0,3,"Farrell, Mr. James",male,40.5,0,0,367232,7.75,,Q
-655,0,3,"Hegarty, Miss. Hanora ""Nora""",female,18.0,0,0,365226,6.75,,Q
-392,1,3,"Jansson, Mr. Carl Olof",male,21.0,0,0,350034,7.7958,,S
-730,0,3,"Ilmakangas, Miss. Pieta Sofia",female,25.0,1,0,STON/O2. 3101271,7.925,,S
-856,1,3,"Aks, Mrs. Sam (Leah Rosen)",female,18.0,0,1,392091,9.35,,S
-32,1,1,"Spencer, Mrs. William Augustus (Marie Eugenie)",female,,1,0,PC 17569,146.5208,B78,C
-207,0,3,"Backstrom, Mr. Karl Alfred",male,32.0,1,0,3101278,15.85,,S
-671,1,2,"Brown, Mrs. Thomas William Solomon (Elizabeth Catherine Ford)",female,40.0,1,1,29750,39.0,,S
-384,1,1,"Holverson, Mrs. Alexander Oskar (Mary Aline Towner)",female,35.0,1,0,113789,52.0,,S
-496,0,3,"Yousseff, Mr. Gerious",male,,0,0,2627,14.4583,,C
-76,0,3,"Moen, Mr. Sigurd Hansen",male,25.0,0,0,348123,7.65,F G73,S
