--- rampwf/hyperopt/tests/interfaces/header_in_files/titanic/requirements.txt
@@ -1,5 +1,0 @@
-numpy
-scikit-learn
-pandas
-matplotlib
-seaborn
